+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* expose the TEA lambda timeout value to allow for DAAC customization
 * add `--platform linux/amd64` to all Docker commands in `Makefile` so `make image` and
 `make container-shell` work on Apple Silicon machines